+++ framework/packages/abstract-std/src/base.rs
@@ -6,7 +6,7 @@ use crate::ibc::{IbcResponseMsg, ModuleIbcMsg};
 // ANCHOR: exec
 /// Wrapper around all possible messages that can be sent to the module.
 #[cosmwasm_schema::cw_serde]
-pub enum ExecuteMsg<BaseMsg, CustomExecMsg, ReceiveMsg = Empty> {
+pub enum ExecuteMsg<BaseMsg, CustomExecMsg> {
     /// A configuration message, defined by the base.
     Base(BaseMsg),
     /// An app request defined by a base consumer.
@@ -18,8 +18,6 @@ pub enum ExecuteMsg<BaseMsg, CustomExecMsg, ReceiveMsg = Empty> {
     /// In order to trust this, the apps and adapters verify this comes from the ibc-host contract.
     /// They should also trust the sending chain
     ModuleIbc(ModuleIbcMsg),
-    /// Receive endpoint for CW20 / external service integrations
-    Receive(ReceiveMsg),
 }
 // ANCHOR_END: exec